Add custom path resolver with react-app-rewired
This commit is adding react-app-rewired to override the webpack configuration without ejecting the app. With this, a new config-overrides.js file is added with a specific path alias '@' to resolve modules from the src directory. Codebase is easier to navigate by reducing relative path deep levels. Also, "react-scripts" commands in package.json are replaced with "react-app-rewired".-based versions. '@mui/material' and '@mui/styled-engine-sc' packages are removed as they are not used anywhere.