Update package.json for Redux implementation and version upgrade
In this commit, the version of the 'turret-build-calculator' package has been upgraded from 0.1.0 to 1.0.0. This indicates a major release, suggesting significant updates and possible breaking changes.

Additionally, '@reduxjs/toolkit' and 'react-redux' have been added to the 'dependencies' section in preparation for the implementation of the Redux library into the project. The 'react-redux' library is used for connecting react components with the redux store. The '@reduxjs/toolkit' is the official and recommended library for setting up redux with a simpler API and additional utilities.

Lastly, 'react-app-rewired' has been moved from 'devDependencies' to 'dependencies'. This package is used to tweak the create-react-app webpack config(s) without using 'eject' command or 'react-scripts'. This change suggests that this package will now be used on the production side of the application rather than only in the development process.

Overall, these changes are key to improving state management and the tweaking of configurations in the application.