Add component uncheck on turret removal
In both src/features/turret-picker/index.tsx and src/features/turret-header/index.tsx, we now handle the checkbox unchecking when turrets are reset or removed. This was achieved by adding a dispatch call to checkboxRemove from "@/reducers/checkbox" for each componentType in the turrets. The unchecking was not performed prior to this change causing unexpected user experience issues.