Add disclaimer text for community ownership
This commit introduces a disclaimer text into the Bottom of the ComponentTable and the GettingStartedOnDesktop pages. A Typography element with the text "Avorion Tools is a community work, and not officially created or maintained by Boxelware" has been added to clarify to users that this is a community project, not an official product of Boxelware. The CardOverflow inside the ComponentTable was also adjusted to have no border radius for a better aesthetic fit.