Rename 'Field' component to 'Numeric' and improve value handling
'Field' component was renamed to 'Numeric' to better reflect its purpose. Changes include altering the props to include a min and max value indicating the range of allowed values an adding additional utility functions to handle NaN values, ensure the number is not less than the minimum and not greater than the maximum. The trimZeroNumber function was removed as the inclusive range and utility functions provide more precision and reliability.