Replace Field component with Numeric in Cargo Add
Field component is replaced with Numeric in the ComponentAddCargo component. This change was made because the Numeric component is more appropriate for dealing with numerical inputs, such as the inputs required in the cargo addition functionality. Now, the max and min quantity of the cargo is validated within Numeric and this helped optimize code by removing unnecessary condition checks.