Update navigation in getting-started page
Added a condition to check whether 'turrets' exists before assessing its length in the 'getting-started' page. This ensures the application does not crash and correctly redirects the user to the 'turret-planner' page if 'turrets' object is present and non-empty. Also, updated the navigation method to replace the current history stack entry with the new location instead of pushing a new entry.