Add transition animation to page changes
Introduced a transition animation when switching between the turret planner page and the getting started page to enhance the user experience. The animation consists of a fade-out/fade-in and a slight scale transition. This was achieved by introducing local state to manage the animation and utilizing the 'clsx' library for merging CSS classes. Furthermore, adjustments were made to the select drop-down within the turret picker component for a better display of options.