Add reset components functionality for turrets
A reset components option has been implemented to allow users to reset all of a turret's components to their defaults. This commit updates the functions in 'turret-header/index.tsx' to include the 'reset-components' action. Localization strings for this option in both English and Russian are added in 'contexts/intl/storage/languages/' directories. A minimum component price constant is also added in 'constants/common.ts'.