Add placeholder for turret selection field
The previous default value of the turret selector has been replaced with a placeholder text. The case-initialized string "select-turret" is used to pull the respective translations from the language files for this placeholder. This change was made to improve the UI by giving clearer instruction to the user while using the turret selection dropdown. New translations have been added to English and Russian language files for the "select-turret" string.