Update turret removal text in both US English and Russian
The language text for removing a turret has been updated for both US English and Russian versions. It has been simplified from "Remove turret" to just "Remove" in both languages. This change is made for better interface usability, as the item being removed will be contextually clear to the user.