Add 'react-viewport-height' to handle screen height
A library called 'react-viewport-height' was added to correctly handle different viewport heights across devices. This was necessary because a solution that relies purely on CSS (e.g. using 'vh') is often inconsistent between devices. In the video background component, this library was used to determine the height of the box containing the video.