Refactor layout and improve code readability
In the turret-planner component, the outermost Box was replaced with a Fragment in order to remove unnecessary DOM node. Additionally, the React imported Fragment was added to clean up the imports.

In the numeric component, multiline style was used to make the decorator props more readable and to adhere to best coding practices.

In the styles module, there were adjustments made in the css file under the .desktop .tableList .table class to improve its responsiveness and provide a cleaner layout. The flex property in the .desktop class was also adjusted to improve the content fitting.

This refactor has improved code readability and the overall layout structure.