Update Numeric component style and Video background image format
Updated the styles.module.css by adding the "user-select: none" rule in the .component section. This is to prevent the contents of the Numeric component from being selected by users accidentally.

For better performance and optimization, the format of the image used in the VideoBackground component has been changed from .png to .webp. The webp format offers better compression and quality characteristics compared to png.