Remove VideoBackground component, update styles
Deleted the 'video-background' component as it was leading to unnecessary complexity and performance issues. This component was removed from turret-planner and getting-started pages and the related styles were also deleted. Adjustments were made to the background styling directly on the body element in 'src/index.css' to maintain consistent background throughout the app. Also, the component CSS classes were renamed to better reflect their role in animations. This change simplifies the codebase, increases rendering performance and maintains the original background aesthetics.