Add max height to turret picker dropdown
The height of the dropdown in the turret picker component was updated to ensure better user experience. The max height of the dropdown list was set to 200px to prevent it from occupying too much screen space when displaying a long list of options. A code comment has also been removed from the global styles.