Add clearComponents functionality to clear turrets action
In the TurretPlannerPage component, the 'clearComponents' action has been imported from the 'component' reducer. This action is then dispatched inside the 'handleClearTurrets' function to ensure that components are also cleared whenever turrets are cleared. This change helps to maintain data consistency by ensuring that component data doesn't linger when related turrets are removed.