Replace Button with IconButton for better UX
Changed the UI implementation from using Buttons to using IconButtons in both turret-planner and cargo-table components for better user experience and visual consistency throughout the application.

In addition, integrated a clear function to reset the cargo in the cargo-table component aimed at improving user's controls and interactivity with the application. Also updated padding in the components-table for better layout presentation.