Enable dynamic text resize for small screens
Enabled dynamic text resizing for smaller screens to enhance the user experience of the application on mobile devices. In the updated files, we are now using a media query to check the screen size and adjust the font size accordingly. 'react-responsive' library was added to the project to enable the use of media queries.

Changes were implemented in the components files and settings in the .idea/inspectionProfiles/Project_Default.xml file were updated to reduce the chance of duplicating the similar code pieces.