Update Typography component width in components-table
In the `src/components/components-table/component-type/index.tsx` file, the width attribute of the Typography component was set to a new value 'max-content'. This was done to ensure the text within the component does not overflow its parent element or obscure any elements that are supposed to be displayed alongside it, promoting better UI structure and user experience.