Update CSS mix-blend-mode and simplify react component
This commit modifies the mix-blend-mode applied to the .icon in two CSS modules from 'difference' to 'exclusion' for consistent rendering across different browsers. The change impacts the 'component-icon' and 'turret-icon' component styles.
In addition, the 'component-type' react component has been simplified to a cleaner syntax by removing the unnecessary Box wrapper around the ComponentIcon. This doesn't affect the functionality but improves readability and performance.