Optimize card display for zero values
This commit introduces an improvement to the UI of the components-table. Specifically, the computation card that was previously always visible, will now only be displayed when there are non-zero computed values (i.e., avg or volume calculations). This should enhance the clarity and cleanliness of the interface, keeping unnecessary information from being displayed when it's not required.