Replace div with Box for better styling support
A div element was replaced with a Box component from '@mui/joy' in the component-type/index.tsx file to allow for better styling support. An 'sx' prop was also added to the Box component to enable the use of the theme-based style function. Additionally, a 'disabled' prop was added to the IconButton component, which dynamically disables the button based on the existence of an entity in the checkbox.entities array.