Add animation to CargoTable component and expose dev command
Added animation to the CargoTable component to provide a smoother transition when items are added or removed. This change was needed to enhance the user experience.

In the package.json file, a "dev:expose" script was added to make the app accessible to network during the development. This was implemented for easier debugging and testing from different devices on the same network.