Add useBreakpoint hook for responsive design
Added a new hook named useBreakpoint to get the current screen size to handle media queries. This hook returns an object indicating whether or not the current viewport matches each breakpoint specified in the application's theme object. This new hook was then used to update checks for screen size in multiple files to improve performance and consistency in handling responsive designs. Also some code formatting and string standardization was done.